+++ Solutions/2015/Day10.cs
@@ -6,14 +6,14 @@ namespace Solutions._2015;
 
 public class Day10 : IDay
 {
-    public int Part1(FileStream fileStream)
+    public long Part1(FileStream fileStream)
     {
         var line = fileStream.ReadSingleLine();
         var result = Solve(line, 40);
         return result;
     }
 
-    public int Part2(FileStream fileStream)
+    public long Part2(FileStream fileStream)
     {
         var line = fileStream.ReadSingleLine();
         var result = Solve(line, 50);
